Express Part:
The App is built on top of Express, which exposes both the API for consumption by the FrontEnd (see below)
or and the native version using Twig Templates.

Angular:

I use Angular 1.6 for this. It sends an http post to the Express endpoints above, gets a response back, and
does the calculations in the angular end.

The app works by simply querying the entire list of guests, returns the searched for voucher in the list, then the template or
frontend does the rest of computations, by making sense out of qos_usage_quota, rx_bytes and tx_bytes.

The calculations are all done in the templates.

Pros:
It works

Cons:
If there're 2 billion guests, the query will return all the guests just to find a SINGLE guest. There's NO way (at least to my knowledge) where
one can query for stats per a voucher, which is so abnormal on the part of Unifi.